lib/driver: put the interrupt handler in RAM
In the makefile of espressif sdk, they add the:

	-ffunction-sections
	-fdata-sections

options in the command-line of GCC. This will tell GCC to put each
function in a separate section. The format of the section name will
be .text.#function name#, without the # (that is, if the function
belongs to the .text section [which by default is true ])

All the section named .text.#function# will be placed in iram or
irom controlled by ld script.

.text.* is placed in .iram1 in eagle.app.v6.4096.ld and it's placed
in .irom0 in esp8266.common.ld

We make sure that all the interrup handler must be in ram.